+++ database/migrations/2020_06_25_112444_create_destockages_table.php
@@ -20,7 +20,7 @@ class CreateDestockagesTable extends Migration
             $table->string('reference')->nullable();
             $table->string('statut')->nullable();
             $table->string('note')->nullable();
-            $table->integer('montant')->nullable();
+            $table->double('montant')->nullable();
             $table->timestamps();
             $table->softDeletes();
